Rename *Test.cs ->Tests.cs
