+++ Support/Key2Joy.Tests/Core/Interop/InteropServerTests.cs
@@ -3,6 +3,6 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace Key2Joy.Tests.Core.Interop;
 
 [TestClass]
-public class InteropServerTest
+public class InteropServerTests
 {
 }
